WGML increment tag usecount only once      allow multiple lines for GML tag attributes      allow numeric labels in macros for ...label and .go      bugfix show_include_stack() if at end of file/macro      additional test file for gml tags
[git-p4: depot-paths = "//depot/openwatcom/": change = 34437]

+++ bld/wgml/h/gvars.h
@@ -74,7 +74,7 @@ global  ulong           line_to;        // ending lineno to process
 #define LINETO_DEFAULT      (0x1000000) // 16 MiB lines should be enough
 
 global  char            gotarget[MAC_NAME_LENGTH +1];   // .go to target name
-global  ulong           gotargetno;     // .go to line no
+global  int32_t         gotargetno;     // .go to line no
 
 global  int             err_count;      // Overall Errorcount
 global  int             wng_count;      // Overall warning count
